// CreateNow/my_app/src/config/types/hotel.ts
export interface NewHotelType {
    cena?: string;
    autor?: string;
    id?: string;
    kategoria?: string;
    miasto?: string;
    nazwa?: string;
    opis?: string;
    telefon?: string;
    zdjecia?: string;
  
  }
  
  export interface AddHotelType {
    cena?: string;
    autor?: string;
    kategoria?: string;
    miasto?: string;
    nazwa?: string;
    opis?: string;
    telefon?: string;
    zdjecia?: string;
    
  
  }